changed CDOTypeImpl.typeID from int to byte

+++ plugins/org.eclipse.emf.cdo.common/src/org/eclipse/emf/cdo/common/model/CDOModelUtil.java
@@ -233,7 +233,7 @@ public final class CDOModelUtil
     return CDOModelUtil.getType(feature.getEType());
   }
 
-  public static CDOType getType(int typeID)
+  public static CDOType getType(byte typeID)
   {
     return CDOTypeImpl.getType(typeID);
   }
